ES | 0-2: Changed from 2rd per. pl to 3rd per. pl
Just because the original sentence in english made a change from using the 2rd person plural to 3rd person plural at the end. That does not mean the same should be done in the translation. The sentence, as it originally was, didn't have a good readability in Spanish.

The proposed change moves to avoid that change by keeping the whole sentence in 2nd person plural. Another solution could be to change the entire sentence to 3rd person plural.